moving test to nvt workflow and moving lambda inside test function
