+++ tests/unit/systems/ops/tf/test_ensemble.py
@@ -16,7 +16,6 @@
 import os
 import shutil
 
-import numpy
 import pytest
 
 os.environ["PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION"] = "python"
@@ -110,64 +109,6 @@ def test_workflow_tf_e2e_config_verification(tmpdir, dataset, engine):
         str(tmpdir), request_schema, df, output_columns, ensemble_config.name
     )
     assert len(response["output"]) == df.shape[0]
-
-
-def raise_(col):
-    if (
-        isinstance(col.dtype, (type(numpy.dtype("float64")), type(numpy.dtype("int64"))))
-        and col.sum() != 0
-    ):
-        return col
-    else:
-        raise ValueError("Number Too High!!")
-
-
-@pytest.mark.skipif(not TRITON_SERVER_PATH, reason="triton server not found")
-@pytest.mark.parametrize("engine", ["parquet"])
-def test_workflow_tf_e2e_error_propagation(tmpdir, dataset, engine):
-    # Create a Workflow
-    schema = dataset.schema
-    for name in ["x", "y", "id"]:
-        dataset.schema.column_schemas[name] = dataset.schema.column_schemas[name].with_tags(
-            [Tags.USER]
-        )
-    selector = ColumnSelector(["x", "y", "id"])
-
-    workflow_ops = selector >> wf_ops.Rename(postfix="_nvt") >> wf_ops.LambdaOp(raise_)
-    workflow = Workflow(workflow_ops["x_nvt"])
-    workflow.fit(dataset)
-
-    # Creating Triton Ensemble
-    triton_chain = selector >> TransformWorkflow(workflow, cats=["x_nvt"])
-    triton_ens = Ensemble(triton_chain, schema)
-
-    # Creating Triton Ensemble Config
-    ensemble_config, node_configs = triton_ens.export(str(tmpdir))
-
-    config_path = tmpdir / ensemble_config.name / "config.pbtxt"
-
-    # Checking Triton Ensemble Config
-    with open(config_path, "rb") as f:
-        config = model_config.ModelConfig()
-        raw_config = f.read()
-        parsed = text_format.Parse(raw_config, config)
-
-        # The config file contents are correct
-        assert parsed.name == "executor_model"
-        assert parsed.platform == "merlin_executor"
-        assert hasattr(parsed, "ensemble_scheduling")
-
-    df = make_df({"x": [0.0, 0.0, 0.0], "y": [4.0, 5.0, 6.0], "id": [7, 8, 9]})
-
-    request_schema = Schema([schema["x"], schema["y"], schema["id"]])
-
-    output_columns = triton_ens.output_schema.column_names
-    with pytest.raises(Exception) as exc:
-        run_ensemble_on_tritonserver(
-            str(tmpdir), request_schema, df, output_columns, ensemble_config.name
-        )
-
-    assert "Number Too High!!" in str(exc.value)
 
 
 @pytest.mark.skipif(not TRITON_SERVER_PATH, reason="triton server not found")
